Make more robust for finding entries
The previous "improvement" actually made it unable to find entries if
there was not space around the equals character. This could still fail
if a line is broken such that it has a single word followed by an equals
sign. This is a real possibility in bib files where the abstract is
broken over lines, although now that Web of Science no longer offers bib
files, it is not so common.